Split testImageOrVideoFormat
