+++ packages/assets/src/detections/parsers/detectAvif.ts
@@ -1,5 +1,5 @@
 import { extensions, ExtensionType } from '@pixi/core';
-import { testImageOrVideoFormat } from '../utils/testImageOrVideoFormat';
+import { testImageFormat } from '../utils/testImageFormat';
 
 import type { FormatDetectionParser } from '..';
 
@@ -13,7 +13,7 @@ export const detectAvif: FormatDetectionParser = {
         // eslint-disable-next-line max-len
         const avifData = 'data:image/avif;base64,AAAAIGZ0eXBhdmlmAAAAAGF2aWZtaWYxbWlhZk1BMUIAAADybWV0YQAAAAAAAAAoaGRscgAAAAAAAAAAcGljdAAAAAAAAAAAAAAAAGxpYmF2aWYAAAAADnBpdG0AAAAAAAEAAAAeaWxvYwAAAABEAAABAAEAAAABAAABGgAAAB0AAAAoaWluZgAAAAAAAQAAABppbmZlAgAAAAABAABhdjAxQ29sb3IAAAAAamlwcnAAAABLaXBjbwAAABRpc3BlAAAAAAAAAAIAAAACAAAAEHBpeGkAAAAAAwgICAAAAAxhdjFDgQ0MAAAAABNjb2xybmNseAACAAIAAYAAAAAXaXBtYQAAAAAAAAABAAEEAQKDBAAAACVtZGF0EgAKCBgANogQEAwgMg8f8D///8WfhwB8+ErK42A=';
 
-        return testImageOrVideoFormat(avifData);
+        return testImageFormat(avifData);
     },
     add: async (formats) => [...formats, 'avif'],
     remove: async (formats) => formats.filter((f) => f !== 'avif'),
